backend/requirements.txt
flask
kafka-python
python-dotenv
pandas